backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
motor
passlib==2.0.1
pydantic
pydantic[email]
